backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==10.4.0
python-decouple==3.8
psycopg[binary]==3.1.19
redis==5.0.7
celery==5.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.2
setuptools<81
dj-database-url==2.1.0

# Professional Document Generation
reportlab==4.2.0
WeasyPrint==62.3
openpyxl==3.1.2
pandas==2.2.2

# Task Scheduling & Automation
django-celery-beat==2.6.0
django-extensions==3.2.3

# Advanced Analytics
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
